Fix qcsg & qrad crashing due to pointer-to-int downconversion

+++ utils/qrad/trace.cpp
@@ -10,6 +10,8 @@
 
 // trace.c
 
+#include <stdint.h>
+
 #include "cmdlib.h"
 #include "mathlib.h"
 #include "bspfile.h"
@@ -75,7 +77,7 @@ void MakeTnodes()
 {
 	// 32 byte align the structs
 	tnodes = reinterpret_cast<tnode_t*>(calloc((numnodes + 1), sizeof(tnode_t)));
-	tnodes = (tnode_t*)(((int)tnodes + 31) & ~31);
+	tnodes = (tnode_t*)(((intptr_t)tnodes + 31) & ~31);
 	tnode_p = tnodes;
 
 	MakeTnode(0);
